refactor(control-plane): group Vercel files under providers/vercel
Move the Vercel sandbox modules added in #700 into a dedicated
src/sandbox/providers/vercel/ directory and drop the redundant
"vercel-" filename prefix:

  providers/vercel-provider.ts -> providers/vercel/provider.ts
  vercel-client.ts             -> providers/vercel/client.ts
  vercel-bootstrap.ts          -> providers/vercel/bootstrap.ts
  vercel-base-snapshot.ts      -> providers/vercel/base-snapshot.ts
  (+ co-located tests)

Keeping the provider under providers/ preserves that directory as the
home for every provider and sets up a per-provider convention that
Modal and Daytona can adopt later. Pure move plus import-path update:
no exported symbols change, so the sandbox barrel's public surface is
identical.
