+++ packages/control-plane/src/routes/repo-images.ts
@@ -16,8 +16,8 @@ import { GlobalSecretsStore } from "../db/global-secrets";
 import { RepoSecretsStore } from "../db/repo-secrets";
 import { mergeSecrets } from "../db/secrets-validation";
 import { createModalClient } from "../sandbox/client";
-import { createVercelSandboxClient } from "../sandbox/vercel-client";
-import { createVercelProvider } from "../sandbox/providers/vercel-provider";
+import { createVercelSandboxClient } from "../sandbox/providers/vercel/client";
+import { createVercelProvider } from "../sandbox/providers/vercel/provider";
 import { resolveSandboxBackendName, supportsRepoImageBackend } from "../sandbox/provider-name";
 import { resolveScmProviderFromEnv } from "../source-control";
 import { createLogger } from "../logger";
